Aplicado regras de acesso

+++ app/Http/Controllers/Admin/UsuarioController.php
@@ -39,11 +39,12 @@ class UsuarioController extends Controller
 
     public function index()
     {
-        /* if(auth()->user()->can('listar-usuarios', false)){
-            dd('sim');
+        if(auth()->user()->can('usuario_listar')){
+            $usuarios = User::all();
+            return view('admin.usuarios.index', compact('usuarios'));
         }else{
-            dd('nao');
-        } */
+            return redirect(route('admin.principal'));
+        }
 
         $usuarios = User::all();
         return view('admin.usuarios.index', compact('usuarios'));
@@ -51,11 +52,19 @@ class UsuarioController extends Controller
 
     public function adicionar()
     {
+        if(!auth()->user()->can('usuario_adicionar')){
+            return redirect(route('admin.principal'));
+        }
+
         return view('admin.usuarios.adicionar');
     }
 
     public function salvar(Request $request)
     {
+        if(!auth()->user()->can('usuario_adicionar')){
+            return redirect(route('admin.principal'));
+        }
+
         $dados = $request->all();
         $usuario = new User();
         $usuario->name = $dados['name'];
@@ -73,12 +82,20 @@ class UsuarioController extends Controller
 
     public function editar($id)
     {
+        if(!auth()->user()->can('usuario_editar')){
+            return redirect(route('admin.principal'));
+        }
+
         $usuario = User::find($id);
         return view('admin.usuarios.editar', compact('usuario'));
     }
 
     public function atualizar(Request $request, $id)
     {
+        if(!auth()->user()->can('usuario_editar')){
+            return redirect(route('admin.principal'));
+        }
+
         $usuario = User::find($id);
         $dados = $request->all();
 
@@ -98,6 +115,10 @@ class UsuarioController extends Controller
 
     public function deletar($id)
     {
+        if(!auth()->user()->can('usuario_deletar')){
+            return redirect(route('admin.principal'));
+        }
+
         User::find($id)->delete();
         Session::flash('mensagem', [
             'msg'=>'Usuário deletado com sucesso!',
@@ -108,6 +129,10 @@ class UsuarioController extends Controller
 
     public function papel($id)
     {
+        if(!auth()->user()->can('usuario_editar')){
+            return redirect(route('admin.principal'));
+        }
+
         $usuario = User::find($id);
         $papel = Papel::all();
 
@@ -116,6 +141,10 @@ class UsuarioController extends Controller
 
     public function salvarPapel(Request $request, $id)
     {
+        if(!auth()->user()->can('usuario_editar')){
+            return redirect(route('admin.principal'));
+        }
+
         $usuario = User::find($id);
         $dados = $request->all();
         $papel = Papel::find($dados['papel_id']);
@@ -126,6 +155,10 @@ class UsuarioController extends Controller
 
     public function removerPapel($id, $papel_id)
     {
+        if(!auth()->user()->can('usuario_editar')){
+            return redirect(route('admin.principal'));
+        }
+
         $usuario = User::find($id);
         $papel = Papel::find($papel_id);
         $usuario->removePapel($papel);
